Added buttons for leaving the simulation and changing the divece in the simulation
